moved caching responsibilities into mcmc instead of being the examples' job (so sampler is still alled if the results are already cached, and then it returns the cached results if they are

+++ examples/common/mcmc.py
@@ -30,38 +30,36 @@ def build_model(resfldr, modelName, model_code):
       sm = pk.load(f)
       f.close()
   return sm
-# def sampler(dnm, lr, datafldr, resfldr, N_samples):
-#   print('STAN: loading data')
-#   X, Y = load_data(os.path.join(datafldr,dnm+'.npz'))
-#   Y[Y == -1] = 0 #convert to Stan LR label style if necessary
 
-#   sampler_data = {'x': X, 'y': Y.astype(int), 'd': X.shape[1], 'n': X.shape[0]}
-
-#   print('STAN: building/loading model')
-#   if lr:
-#     sm = build_model(resfldr, 'pystan_model_logistic.pk', logistic_code)
-#   else:
-#     sm = build_model(resfldr, 'pystan_model_poisson.pk', poisson_code)
 def sampler(dnm, datafldr, resfldr, N_samples, stan_representation):
-  print('STAN: loading data')
-  X, Y = load_data(os.path.join(datafldr,dnm+'.npz'))
-  Y[Y == -1] = 0 #convert to Stan LR label style if necessary
 
-  sampler_data = {'x': X, 'y': Y.astype(int), 'd': X.shape[1], 'n': X.shape[0]}
+  if not os.path.exists('caching/'):
+  os.mkdir('caching')
 
-  print('STAN: building/loading model')
-  name, code = stan_representation
-  sm = build_model(resfldr, name, code)
+  if os.path.exists('caching/'+dnm+'_samples.npy'):
+    print("Using cached samples for " + dnm)
+    return np.load('caching/'+dnm+'_samples.npy')
+  else:
+    print('No MCMC samples found -- running STAN')
+    print('STAN: loading data')
+    X, Y = load_data(os.path.join(datafldr,dnm+'.npz'))
+    Y[Y == -1] = 0 #convert to Stan LR label style if necessary
 
-  #'pystan_model_logistic.pk'
-  #'pystan_model_poisson.pk'
-  #logistic_code
+    sampler_data = {'x': X, 'y': Y.astype(int), 'd': X.shape[1], 'n': X.shape[0]}
 
-  print('STAN: sampling posterior: ' + dnm)
-  t0 = time.process_time()
-  thd = sampler_data['d']+1
-  fit = sm.sampling(data=sampler_data, iter=N_samples*2, chains=1, control={'adapt_delta':0.9, 'max_treedepth':15}, verbose=True)
-  samples = fit.extract(permuted=False)[:, 0, :thd]
-  np.save(os.path.join(resfldr, dnm+'_samples.npy'), samples) 
-  tf = time.process_time()
-  np.save(os.path.join(resfldr, dnm+'_mcmc_time.npy'), tf-t0)
+    print('STAN: building/loading model')
+    name, code = stan_representation
+    sm = build_model(resfldr, name, code)
+
+    print('STAN: sampling posterior: ' + dnm)
+    t0 = time.process_time()
+    thd = sampler_data['d']+1
+    fit = sm.sampling(data=sampler_data, iter=N_samples*2, chains=1, control={'adapt_delta':0.9, 'max_treedepth':15}, verbose=True)
+    samples = fit.extract(permuted=False)[:, 0, :thd]
+    np.save(os.path.join(resfldr, dnm+'_samples.npy'), samples) 
+    tf = time.process_time()
+    np.save(os.path.join(resfldr, dnm+'_mcmc_time.npy'), tf-t0)
+    return samples
+
+  #if not cached, run sampler.
+  #else return with message "using cached data" - this is definitely something mcmc should do, not individual examples
